Refactor: Move methods into another classes
