+++ app/src/androidTest/java/ch/epfl/sweng/SDP/game/VotingPageActivityTest.java
@@ -50,6 +50,7 @@ import static org.mockito.Mockito.when;
 public class VotingPageActivityTest {
 
     private static final String USER_ID = "userA";
+    private static final String ROOM_ID_TEST = "0123457890";
 
     private DataSnapshot dataSnapshotMock;
     private DatabaseError databaseErrorMock;
@@ -71,7 +72,7 @@ public class VotingPageActivityTest {
                 @Override
                 protected Intent getActivityIntent() {
                     Intent intent = new Intent();
-                    intent.putExtra("RoomID", "0123457890");
+                    intent.putExtra("RoomID", ROOM_ID_TEST);
                     return intent;
                 }
             };
@@ -93,7 +94,7 @@ public class VotingPageActivityTest {
     public void ratingUsingRatingBarShouldBeSaved() {
 
         // To ensure that the rating value does not get above 20
-        Database.getReference("realRooms.0123457890.ranking." + USER_ID).setValue(0);
+        Database.getReference("realRooms." + ROOM_ID_TEST + ".ranking." + USER_ID).setValue(0);
 
         short counter = mActivityRule.getActivity().getChangeDrawingCounter();
         SystemClock.sleep(5000);
@@ -163,8 +164,8 @@ public class VotingPageActivityTest {
         SystemClock.sleep(2000);
         intended(hasComponent(HomeActivity.class.getName()));
         Intents.release();
-        Database.getReference("realRooms.0123457890.users." + USER_ID).setValue(USER_ID);
-        Database.getReference("realRooms.0123457890.ranking." + USER_ID).setValue(0);
+        Database.getReference("realRooms." + ROOM_ID_TEST + ".users." + USER_ID).setValue(USER_ID);
+        Database.getReference("realRooms." + ROOM_ID_TEST + ".ranking." + USER_ID).setValue(0);
         SystemClock.sleep(2000);
     }
