+++ app/src/androidTest/java/ch/epfl/sweng/SDP/utils/NetworkStateReceiverTest.java
@@ -4,6 +4,7 @@ import android.support.test.InstrumentationRegistry;
 
 import org.junit.Test;
 
+import ch.epfl.sweng.SDP.utils.network.ConnectivityWrapper;
 import ch.epfl.sweng.SDP.utils.network.NetworkStateReceiver;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -12,6 +13,7 @@ import static org.hamcrest.Matchers.is;
 public class NetworkStateReceiverTest {
     @Test
     public void testInternetConnection() {
-        assertThat(NetworkStateReceiver.isOnline(InstrumentationRegistry.getContext()), is(true));
+        assertThat(ConnectivityWrapper.isOnline(InstrumentationRegistry.getContext()),
+                                        is(true));
     }
 }
